+++ bridge/bridge.go
@@ -177,7 +177,7 @@ func (b *Bridge) add(containerId string, quiet bool) {
 func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
 	container := port.container
 	defaultName := strings.Split(path.Base(container.Config.Image), ":")[0]
-	
+
 	// not sure about this logic. kind of want to remove it.
 	hostname, err := os.Hostname()
 	if err != nil {
@@ -207,10 +207,10 @@ func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
 	service.ID = hostname + ":" + container.Name[1:] + ":" + port.ExposedPort
 	service.Name = mapDefault(metadata, "name", defaultName)
 	if isgroup {
-		 service.Name += "-" + port.ExposedPort
+		service.Name += "-" + port.ExposedPort
 	}
 	var p int
-	if b.config.Internal == true {
+	if b.isInternal(port) == true {
 		service.IP = port.ExposedIP
 		p, _ = strconv.Atoi(port.ExposedPort)
 	} else {
@@ -282,4 +282,16 @@ func (b *Bridge) didExitCleanly(containerId string) bool {
 		return false
 	}
 	return !container.State.Running && container.State.ExitCode == 0
+}
+
+func (b *Bridge) isInternal(port ServicePort) bool {
+	container := port.container
+
+	for _, e := range container.Config.Env {
+		if ok, key, value := parseEnv(e); ok && key == "internal" {
+			return value == "true"
+		}
+	}
+
+	return b.config.Internal
 }